form-ocr/requirements.txt
fastapi
uvicorn
pytesseract
opencv-python-headless
numpy
Pillow
pdf2image
openpyxl
fpdf
python-multipart